+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the `C-mantic` extension will be documented in this file.
 ### Added
 - Added support for `s_` static member naming scheme when generating getters and setters.
 
+### Fixed
+- Fixed a bug where classes/structs defined within a class/struct could affect the placement of getters and setters.
+
 ## [0.4.0] - February 14, 2021
 ### Added
 - Added `cmantic.moveDefinitionIntoOrOutOfClass` command/code-action. Additionally, definitions within classes can be moved to a matching source file. (#7)